docs(README): Update badges
Updated the link for the version badge to point to the npm page for all of the angular packages.
Removed the downloads badge.

The downloads badge was only showing the number of angular2 beta downloads, and not counting any RC downloads. The npm downloads api doesn't currently support scoped packages so we're unable to use @angular/core for the downloads badge like we are for the version badge. Support is being worked on here: https://github.com/npm/download-counts/issues/23